services/worker/requirements.txt
pika==1.3.2
httpx==0.27.0
python-dotenv==1.0.1
PyGithub==2.3.0
groq==0.9.0
PyJWT==2.8.0
cryptography==42.0.5
tenacity==8.2.3